initial support for multiple Discovery indexes
