+++ packages/frontend/src/pages/Discovery/data.ts
@@ -8,7 +8,6 @@ import type { NavPageLayoutProps } from '../../features/Navigation';
 export const DiscoveryPageGetServerSideProps: GetServerSideProps<
   NavPageLayoutProps
 > = async () => {
-
   try {
     const discoveryConfig: DiscoveryProps = await ContentSource.get(
       `config/${GEN3_COMMONS_NAME}/discovery.json`,
